refactor: split ai_engine.py into separate services for different ML functionalities
